requirements.txt
# Updated dependencies to address security vulnerabilities found by pip-audit
fastapi==0.109.1
uvicorn==0.23.2
jinja2==3.1.6
python-dotenv==1.0.0
python-multipart==0.0.18
twilio==8.9.1
requests==2.32.4
httpx==0.25.0
sqlalchemy==2.0.43
alembic==1.12.1
redis==5.0.1
pydantic==2.12.0a1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
groq==0.4.1
phonenumbers==8.13.25
gunicorn==23.0.0
email-validator
black==24.4.2
flake8==7.0.0
pytest==8.1.1
pytest-cov==5.0.0
playwright==1.44.0
selenium==4.21.0
stripe==9.3.0
reportlab==4.2.2
sentry-sdk[fastapi]==1.38.0
psutil==5.9.8
psycopg2-binary==2.9.9